+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/gateway
 
+## 3.0.124
+
+### Patch Changes
+
+- 286b7a2: Backport: chore(provider/gateway): update gateway model settings files
+
+## 3.0.123
+
+### Patch Changes
+
+- 537a022: Backport: chore(provider/gateway): update gateway model settings files
+
+## 3.0.122
+
+### Patch Changes
+
+- 9766034: Backport: chore(provider/gateway): update gateway model settings files
+
 ## 3.0.121
 
 ### Patch Changes