+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/langchain
 
+## 2.0.203
+
+### Patch Changes
+
+- ai@6.0.196
+
+## 2.0.202
+
+### Patch Changes
+
+- ai@6.0.195
+
+## 2.0.201
+
+### Patch Changes
+
+- ai@6.0.194
+
 ## 2.0.200
 
 ### Patch Changes