+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/svelte
 
+## 4.0.196
+
+### Patch Changes
+
+- ai@6.0.196
+
+## 4.0.195
+
+### Patch Changes
+
+- ai@6.0.195
+
+## 4.0.194
+
+### Patch Changes
+
+- ai@6.0.194
+
 ## 4.0.193
 
 ### Patch Changes