+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/vue
 
+## 3.0.196
+
+### Patch Changes
+
+- ai@6.0.196
+
+## 3.0.195
+
+### Patch Changes
+
+- ai@6.0.195
+
+## 3.0.194
+
+### Patch Changes
+
+- ai@6.0.194
+
 ## 3.0.193
 
 ### Patch Changes